extract ProgressBar
